#2 I think I found something...
I think i forgot the increment  the variable pcount ... this has been
corrected.
There may be a new issue working its way into the code. It is something
that has happened only with the addition of the two other coded
functions each post gets sent 2x times, and so the second time fails
because there is alredy a post there with that name... ill look into
that here too.