delete posts , including Purges works and branc will be merged
What we have done is given an UI to the admin staff to undelete and
purge spam and delete flagged messages from the database. This has now
been tested in live and it works. Please also note that we made some
other changes about, we also added the ability of a user to flag a post
as deleted , there is a bug but that will be dealt wit in an issue.