Enhancement
Added code to make it obvious what messages were actually unread... This
will be how you mark messages as being read. rather than the button in
the message field. It should be noted that this button didnt always
works as expected so This enhancement now works better. ALso fixed an
issue with the showing top message when opened... it was refusing to
open other messages... now all messages work as expected.