Enhancement
Added the ability to add preformatted text e.g. quotes and code
segments. To do this in messages or posts youll just need to include an
open and close <pre> tag and the formatting type that you wish to use.
<code> or <quote> open and closeing must be in the <pre></pre> tags.
Now displayse line breaks cleanly, this allows you to depict paragraphs
correctly when they are read back into the message system. just use your
enter key to set a line break.

New column added to the IssueMessage table along with a foreign key,
this will allow you to  search using the issue number.