Enhancement
This new enhancement for Tables.php now has two columns that will allow
an admin to require posted content (posts, comments, content) to be
approved by a admin or mod, to set as a default option swap the
#contentType#ModApprovalRequired default val from 0 to 1, set by default
as 0.
Started the code for the issue tracker system, Note that this will
basiclly be the mod que, as well as the issue tracker.