Issue #9 :: more steps to corection being completed
Okay I have just about fixed all the remaining issues with compatibility
between MySQL and MariaDB, that were showing up, the fact now is just
one more test yto ensure that the code again is truely grabbing the
correct array variable. The process was changed as I was running into
issues with the Show query was returning slightly different results, So
now it searches through the array for Version_comment and then puts the
Value into the php variable. This seems to be working correctly and
again just needs a bit further testing.