Updated ReadMe with Current Project status
Will be moving over to class and OOP based controllers, which will make things like my 600+ line code for the creation of the tables completely beautified and shorted in code than current. Will also hopefully reduce over all code on the system and thus more efficient. I ask that you bear with me if you are following the progress of this project.